pyproject.toml
[project]
name = "benchmax"
version = "0.1.2"
description = "Framework-Agnostic RL Environments for LLM Fine-Tuning"
readme = "README.md"
authors = [{ name = "cgft.io" }]
requires-python = "==3.12.*"
dependencies = [
    "aiohttp>=3.13.1",
    "asyncio>=4.0.0",
    "datasets>=4.0.0",
    "fastmcp~=2.12.0",
    "pyjwt>=2.10.1",
    "skypilot~=0.8.1",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.2",
]
skypilot = [
    "skypilot[aws,gcp,azure]~=0.8.1", # Change this to your cloud provider
    "pip>=25.3",                      # Added as needed for skypilot launch
    "msrestazure>=0.6.4.post1",
]
skyrl = [
    "grpcio>=1.60.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "ray>=2.48.0",
    "skyrl-gym>=0.1.1",
    "skyrl-train[vllm]>=0.2.0",
]
excel = ["openpyxl>=3.1.5"]
excel-mac-windows = ["openpyxl>=3.1.5", "xlwings>=0.33.16"]
crm = ["python-dateutil>=2.9.0.post0", "simple-salesforce>=1.12.9"]

[tool.uv]
conflicts = [[{ group = "skypilot" }, { group = "skyrl" }]]

[tool.uv.pip]
extra = ["dev", "skypilot", "skyrl", "excel", "excel-mac-windows", "crm"]

[tool.uv.extra-build-dependencies]
flash-attn = [{ requirement = "torch", match-runtime = true }]

[tool.uv.extra-build-variables]
flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "TRUE" }
